+++ pom.xml
@@ -57,6 +57,18 @@
             <artifactId>logback-slack-appender</artifactId>
             <version>1.4.0</version>
         </dependency>
+        <!-- 파일 업로드, 다운로드-->
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+            <version>3.0</version>
+        </dependency>
+
+        <dependency>
+            <groupId>commons-io</groupId>
+            <artifactId>commons-io</artifactId>
+            <version>2.5</version>
+        </dependency>
     </dependencies>
 
     <build>
