--- src/main/java/com/hoin/boardStudy/BoardApplication.java
@@ -1,13 +1,0 @@
-package com.hoin.boardStudy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BoardApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BoardApplication.class, args);
-    }
-
-}
