--- src/main/java/com/hoin/boardStudy/board/config/WebMvcConfigure.java
@@ -1,20 +1,0 @@
-package com.hoin.boardStudy.board.config;
-
-import com.hoin.boardStudy.board.config.handler.PageRequestHandler;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-public class WebMvcConfigure implements WebMvcConfigurer {
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(PageRequestHandler());
-    }
-
-    public PageRequestHandler PageRequestHandler() {
-        return new PageRequestHandler();
-    }
-}
