--- src/main/java/com/hoin/boardStudy/board/dto/PageInfo.java
@@ -1,13 +1,0 @@
-package com.hoin.boardStudy.board.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PageInfo {
-    private int page;
-    private int pageSize;
-}
