+++ src/main/java/com/hoin/boardstudy/BoardApplication.java
@@ -1,0 +1,13 @@
+package com.hoin.boardstudy;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BoardApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BoardApplication.class, args);
+    }
+
+}
