+++ src/main/java/com/hoin/boardstudy/board/controller/TestC.java
@@ -1,0 +1,16 @@
+package com.hoin.boardstudy.board.controller;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+@Slf4j
+public class TestC {
+    
+    @GetMapping("/test")
+    public void test() {
+        log.error("에러");
+        throw new RuntimeException("에러쓰");
+    }
+}
