+++ src/main/java/com/hoin/boardstudy/board/dto/BoardSaveRequest.java
@@ -1,0 +1,20 @@
+package com.hoin.boardstudy.board.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class BoardSaveRequest {
+    private int boardId; // 글 번호
+    private String writer; // 작성자
+    private String title; // 제목
+    private String content; // 내용
+    private LocalDateTime currentTime = LocalDateTime.now(); // 현재시간
+}
