+++ src/main/java/com/hoin/boardstudy/board/dto/PageInfo.java
@@ -1,0 +1,13 @@
+package com.hoin.boardstudy.board.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class PageInfo {
+    private int page;
+    private int pageSize;
+}
