+++ src/main/java/com/hoin/boardstudy/user/service/LoginVerification.java
@@ -1,0 +1,28 @@
+package com.hoin.boardstudy.user.service;
+
+import com.hoin.boardstudy.user.dto.User;
+import com.hoin.boardstudy.user.mapper.UserMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LoginVerification {
+
+    public final PasswordManagement passwordManagement;
+    public final UserService userService;
+    public final UserMapper userMapper;
+    BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+
+    /* 로그인 검증 */
+    public boolean loginVerification(User user, String rawPassword) {
+        user = userService.getUserInfo(user.getUserId());
+        if (user != null &&
+                passwordEncoder.matches(rawPassword, user.getPassword())) {
+            return true;
+        }
+            return false;
+    }
+
+}
