+++ src/main/resources/mybatis/mapper/UserMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
-<mapper namespace="com.hoin.boardStudy.user.mapper.UserMapper">
+<mapper namespace="com.hoin.boardstudy.user.mapper.UserMapper">
 
     <resultMap id="userMap" type="user">
         <result column="user_id" property="userId"/>
