+++ src/main/resources/templates/board/writeForm.html
@@ -12,12 +12,26 @@
 
     <div class="container" style="margin-top: 100px">
         <h2>글쓰기</h2>
-        <form name="writeForm" id="writeForm" role="form" method="post" action="saveBoard.do"/>
+        <form name="writeForm" id="writeForm" role="form" method="post" enctype="multipart/form-data" action="saveBoard.do"/>
             <table class="table table-borderless" >
                 <tr><th>제목</th></tr>
                 <tr><td width="90%"><input name="title" id="title" class="form-control" placeholder="제목을 입력하세요." required="required" ></td></tr>
                 <tr><th>내용</th></tr>
                 <tr><td class="text-center"><textarea class="form-control" name="content" id="content" maxlength="1500" rows="20" style="resize: none; background-color:white" placeholder="내용을 입력하세요." required="required"></textarea></td></tr>
+                <tr><th>파일 첨부</th></tr>
+                <tr>
+                    <td>
+                        <div class="input-group">
+                            <div class="custom-file">
+                                <input type="file" class="custom-file-input" name="file" id="file">
+                                <label class="custom-file-label" for="file">Choose file</label>
+                            </div>
+                            <div class="input-group-append">
+                                <button class="btn btn-outline-secondary" type="button">Button</button>
+                            </div>
+                        </div>
+                    </td>
+                </tr>
             </table>
             <div  class="d-grid gap-2 d-md-flex justify-content-md-end">
                 <button type="button" class="btn btn-secondary" onclick="location.href='list.do'">목록</button>
