--- src/test/java/com/hoin/boardStudy/BoardApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hoin.boardStudy;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BoardApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
